fix types in dtos

+++ ProsthesisEmulator/src/main/kotlin/emulator/dto/gesture/GestureActionDto.kt
@@ -4,10 +4,10 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class GestureActionDto(
-        val pointerFingerPosition: Byte = 0,
-        val middleFingerPosition: Byte = 0,
-        val ringFinderPosition: Byte = 0,
-        val littleFingerPosition: Byte = 0,
-        val thumbFingerPosition: Byte = 0,
-        val delay: Short = 0
+        val pointerFingerPosition: Int = 0,
+        val middleFingerPosition: Int = 0,
+        val ringFinderPosition: Int = 0,
+        val littleFingerPosition: Int = 0,
+        val thumbFingerPosition: Int = 0,
+        val delay: Int = 0
 )
